Move ITestConfigSettingsService up one directory.
