+++ src/client/unittests/common/services/configSettingService.ts
@@ -3,7 +3,8 @@ import { Uri, WorkspaceConfiguration } from 'vscode';
 import { IWorkspaceService } from '../../../common/application/types';
 import { Product } from '../../../common/types';
 import { IServiceContainer } from '../../../ioc/types';
-import { ITestConfigSettingsService, UnitTestProduct } from './../types';
+import { ITestConfigSettingsService } from '../../types';
+import { UnitTestProduct } from './../types';
 
 @injectable()
 export class TestConfigSettingsService implements ITestConfigSettingsService {
