+++ src/client/unittests/types.ts
@@ -56,6 +56,13 @@ export interface IUnitTestManagementService {
     viewOutput(cmdSource: CommandSource): void;
 }
 
+export const ITestConfigSettingsService = Symbol('ITestConfigSettingsService');
+export interface ITestConfigSettingsService {
+    updateTestArgs(testDirectory: string | Uri, product: UnitTestProduct, args: string[]): Promise<void>;
+    enable(testDirectory: string | Uri, product: UnitTestProduct): Promise<void>;
+    disable(testDirectory: string | Uri, product: UnitTestProduct): Promise<void>;
+}
+
 export interface ITestConfigurationManager {
     requiresUserToConfigure(wkspace: Uri): Promise<boolean>;
     configure(wkspace: Uri): Promise<void>;
